+++ CHANGELOG.md
@@ -26,3 +26,8 @@
 ## [1.2.3]
 
 - Fix scanner compatibility with asdf CLI manager
+- Remove deprecate install.sh usage
+
+## [1.2.4]
+
+- Update MCP Setup to use current XYGENI_TOKEN